replace f64 with Decimal for all financial values
Bid.price, Impression.bidfloor, AdEvent.clearing_price, CampaignMetrics.spend,
and all price query params across services. OpenRTB fields use serde float
serialization for JSON number compatibility. Drops the FLOAT8::NUMERIC cast in
the reconciliation SQL query.

+++ libs/core-models/src/lib.rs
@@ -1,3 +1,4 @@
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 
 // ==========================================
@@ -15,7 +16,8 @@ pub struct BidRequest {
 #[derive(Debug, Deserialize, Serialize)]
 pub struct Impression {
     pub id: String,
-    pub bidfloor: Option<f64>,
+    #[serde(with = "rust_decimal::serde::float_option")]
+    pub bidfloor: Option<Decimal>,
 }
 
 #[derive(Debug, Deserialize, Serialize)]
@@ -34,7 +36,7 @@ pub struct Device {
 #[derive(Debug, Deserialize, Serialize)]
 pub struct User {
     pub id: Option<String>,
-    pub buyeruid: Option<String>, // ADDED: Web DSP Cookie ID
+    pub buyeruid: Option<String>,
 }
 
 // ==========================================
@@ -56,20 +58,21 @@ pub struct SeatBid {
 pub struct Bid {
     pub id: String,
     pub impid: String,
-    pub price: f64,
+    #[serde(with = "rust_decimal::serde::float")]
+    pub price: Decimal,
     pub adid: Option<String>,
     pub crid: Option<String>,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct AdEvent {
-    pub event_type: EventType, // e.g., "IMPRESSION" or "CLICK"
+    pub event_type: EventType,
     pub campaign_id: String,
     pub bid_id: String,
-    pub clearing_price: Option<f64>,
+    #[serde(with = "rust_decimal::serde::float_option")]
+    pub clearing_price: Option<Decimal>,
     pub timestamp_ms: u128,
 }
-
 
 #[derive(Debug, Serialize, Deserialize, PartialEq, Eq)]
 #[serde(rename_all = "UPPERCASE")]
